requirements-dev.txt
django-nose>=1.4.3,<1.5
